PointsService: Database size reduction (BIGINT to INT)

+++ src/PointsService/Core/Entity/DailyStat.cs
@@ -11,6 +11,6 @@ public class DailyStat
     public string UserId { get; set; } = null!;
 
     public DateOnly Date { get; set; }
-    public long MessagePoints { get; set; }
-    public long ReactionPoints { get; set; }
+    public int MessagePoints { get; set; }
+    public int ReactionPoints { get; set; }
 }
